requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
supabase==2.21.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
psycopg[binary]==3.1.12
httpx==0.28.1
requests>=2.28.0
